LayoutTests/fast/speechsynthesis/speech-synthesis-empty-initial-voice-list-expected.txt

This tests that the voice list can initially be empty and populated after voiceschanged event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS initialVoices.length is 0
PASS voices.length > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
